Add option to disable asset publishing
This update adds a new static property and method in Horizon.php to allow disabling asset publishing. The method affects the asset validation in `assetsAreCurrent()` and the returned view data in `HomeController`. Theme-dependent inline styles and scripts in layout.blade.php were updated according to the new option as well.